rest controller befejezve

+++ hr/src/main/java/hu/webuni/hr/andro/web/HrController.java
@@ -19,9 +19,9 @@ public class HrController {
 	private List<Employee> employees = new ArrayList<>();
 	
 	{
-		employees.add(new Employee((long) 723, "Teszt Elek", "rendszergazda", 450000, LocalDateTime.of(2019, 10, 2, 0, 0)));
-		employees.add(new Employee((long) 561, "Nap Pali", "grafikus", 600000, LocalDateTime.of(2015, 4, 12, 0, 0)));
-		employees.add(new Employee((long) 278, "Tra Pista", "rendszer tervező", 800000, LocalDateTime.of(2011, 2, 23, 0, 0)));
+		employees.add(new Employee(723L, "Teszt Elek", "rendszergazda", 450000, LocalDateTime.of(2019, 10, 2, 0, 0)));
+		employees.add(new Employee(561L, "Nap Pali", "grafikus", 600000, LocalDateTime.of(2015, 4, 12, 0, 0)));
+		employees.add(new Employee(278L, "Tra Pista", "rendszer tervező", 800000, LocalDateTime.of(2011, 2, 23, 0, 0)));
 	}
 	
 	private Employee getEmployee(long id) {
